fix: Устранение замечаний, кроме instanceof
