+++ app.js
@@ -3,8 +3,9 @@ const express = require('express');
 
 const app = express();
 const mongoose = require('mongoose');
+const helmet = require('helmet');
 const { errors } = require('celebrate');
-const { PORT, MONGOOSE_URL } = require('./devConfig');
+const { PORT, MONGOOSE_URL } = require('./config');
 
 const iternalServerError = require('./errors/iternalServerError');
 const { NOT_FOUND_ERROR } = require('./errors/notFoundError');
@@ -14,6 +15,7 @@ const { cors } = require('./middlewares/cors');
 const { limiter } = require('./utils/rateLimit');
 const { requestLogger, errorLogger } = require('./middlewares/logger');
 
+app.use(helmet());
 app.use(express.json());
 app.use(express.urlencoded({ extended: true }));
 app.use(requestLogger);
