+++ controllers/user.js
@@ -1,9 +1,9 @@
 const jwt = require('jsonwebtoken');
 const bcrypt = require('bcryptjs');
 const User = require('../models/user');
-const { errorMessage } = require('../utils/errorMessage');
+const { handlerMessageError } = require('../utils/handlerMessageError');
 const { NOT_FOUND_ERROR } = require('../errors/notFoundError');
-const { JWT_STORAGE_TIME, SALT_LENGTH, JWT_SECRET } = require('../devConfig');
+const { JWT_STORAGE_TIME, SALT_LENGTH, JWT_SECRET } = require('../config');
 const { CREATED, OK } = require('../utils/successes');
 
 // получить текущего пользователя
@@ -13,7 +13,7 @@ const getUser = (req, res, next) => {
       throw new NOT_FOUND_ERROR('Пользователь с таким id не найден');
     })
     .then((user) => res.status(OK).send(user))
-    .catch((err) => errorMessage(err, req, res, next));
+    .catch((err) => handlerMessageError(err, req, res, next));
 };
 // обновить текущего пользователя
 const patchUser = (req, res, next) => {
@@ -29,22 +29,21 @@ const patchUser = (req, res, next) => {
       throw new NOT_FOUND_ERROR('Пользователь с таким id не найден');
     })
     .then((user) => res.send(user))
-    .catch((err) => errorMessage(err, req, res, next));
+    .catch((err) => handlerMessageError(err, req, res, next));
 };
 // создать пользователя
 const createUser = (req, res, next) => {
   const { name, email, password } = req.body;
 
   bcrypt
-    .hash(password, process.env.SALT_LENGTH || SALT_LENGTH)
+    .hash(password, SALT_LENGTH || process.env.SALT_LENGTH)
     .then((hash) => User.create({
       name,
       email,
       password: hash,
     }))
-    .then((user) => User.findOne({ _id: user._id }))
     .then((user) => res.status(CREATED).send({ data: user }))
-    .catch((err) => errorMessage(err, req, res, next));
+    .catch((err) => handlerMessageError(err, req, res, next));
 };
 // авторизация пользователя
 const login = (req, res, next) => {
