+++ errors/badRequestError.js
@@ -1,7 +1,6 @@
 class BAD_REQUEST_ERROR extends Error {
   constructor(message) {
     super(message);
-    this.errorMessage = message;
     this.statusCode = 400;
   }
 }
