+++ errors/forbiddenError.js
@@ -1,7 +1,6 @@
 class FORBIDDEN_ERROR extends Error {
   constructor(message) {
     super(message);
-    this.errorMessage = message;
     this.statusCode = 403;
   }
 }
