+++ errors/notFoundError.js
@@ -1,7 +1,6 @@
 class NOT_FOUND_ERROR extends Error {
   constructor(message) {
     super(message);
-    this.errorMessage = message;
     this.statusCode = 404;
   }
 }
