+++ middlewares/cors.js
@@ -1,9 +1,4 @@
-const allowedCors = [
-  'https://movies.explorer.nomoredomains.rocks',
-  'http://movies.explorer.nomoredomains.rocks',
-  'http://localhost:3000',
-  'https://localhost:3000',
-];
+const allowedCors = require('../config');
 
 const cors = (req, res, next) => {
   const { origin } = req.headers;
