+++ package.json
@@ -29,6 +29,7 @@
     "express": "^4.18.1",
     "express-rate-limit": "^6.7.0",
     "express-winston": "^4.2.0",
+    "helmet": "^7.0.0",
     "jsonwebtoken": "^9.0.0",
     "mongoose": "^7.2.0",
     "validator": "^13.9.0",
